Fix events when there's a span inside a button
A button normally only allows text inside of it. If we have a span
inside (like is the case with the ones that use Material Icons symbols),
it intercepts all pointer events, thus messing with the event target
when we are accessing it in the JS code.

Therefore the solution is to disable the pointer events for the span,
forcing the browser to foward those to the button.

See: https://stackoverflow.com/questions/26402033/missing-click-event-for-span-inside-button-element-on-firefox